separate code into src and tests directories
