+++ backend/Dockerfile
@@ -2,18 +2,20 @@
 FROM python:3.12
 
 # Set up a working directory
-WORKDIR /backend_code
+WORKDIR /backend
 
 # Copy just the requirements into the working directory so it gets cached by itself
-COPY requirements.txt /backend_code/requirements.txt
+COPY requirements.txt /backend/requirements.txt
 
 # Install the dependencies from the requirements file
-RUN pip install --no-cache-dir --upgrade -r /backend_code/requirements.txt
+RUN pip install --no-cache-dir --upgrade -r /backend/requirements.txt
 
-# Copy the code into the working directory
-COPY . /backend_code
+# Copy the source code into the working directory
+COPY ./src/. /backend/src
 
 EXPOSE 8250
 
+RUN ls -la
+
 # Tell uvicorn to start spin up our code, which will be running inside the container now
-CMD ["uvicorn", "api:app", "--host", "0.0.0.0", "--port", "8250"]
+CMD ["uvicorn", "src.api:app", "--host", "0.0.0.0", "--port", "8250"]
