+++ backend/src/agents/adapters.py
