+++ backend/src/prompts/__init__.py
@@ -1,3 +1,3 @@
 from .prompting import PromptEngine
 
-__all__ = [ "PromptEngine" ]
+__all__ = ["PromptEngine"]
