+++ backend/tests/agents/adapters_test.py
@@ -1,6 +1,6 @@
-from agents import tool_metadata
-from .adapters import convert_to_mistral_tool
-from .types import Parameter
+from src.agents import tool_metadata
+from src.agents.adapters import convert_to_mistral_tool
+from src.agents.types import Parameter
 
 name = "Mock Tool"
 description = "A test tool"
