+++ backend/tests/utils/llm_test.py
@@ -1,10 +1,9 @@
-from utils.llm import call_model, call_model_with_tools, MistralClient
 from unittest.mock import MagicMock
 from mistralai.models.chat_completion import ChatCompletionResponse
 from mistralai.models.chat_completion import ChatCompletionResponseChoice
 from mistralai.models.chat_completion import ChatMessage
 from mistralai.models.common import UsageInfo
-from utils import Config
+from src.utils import Config, call_model, call_model_with_tools, MistralClient
 
 system_prompt = "system_prompt"
 user_prompt = "user_prompt"
@@ -25,7 +24,7 @@ mock_config = MagicMock(spec=Config)
 
 
 def test_call_model_returns_string(mocker):
-    client_instance = mocker.patch("utils.llm.client", return_value=mock_client)
+    client_instance = mocker.patch("src.utils.llm.client", return_value=mock_client)
     client_instance.chat.return_value = create_mock_chat_response(content_response)
 
     response = call_model(system_prompt, user_prompt)
@@ -34,8 +33,8 @@ def test_call_model_returns_string(mocker):
 
 
 def test_call_model_calls_client_chat(mocker):
-    config_instance = mocker.patch("utils.llm.config", return_value=mock_config)
-    client_instance = mocker.patch("utils.llm.client", return_value=mock_client)
+    config_instance = mocker.patch("src.utils.llm.config", return_value=mock_config)
+    client_instance = mocker.patch("src.utils.llm.client", return_value=mock_client)
 
     call_model(system_prompt, user_prompt)
 
@@ -72,7 +71,7 @@ tool_calls = [{"id": "1", "type": "function", "function": {"name": "mock_tool", 
 
 
 def test_call_model_with_tools_returns_tuple(mocker):
-    client_instance = mocker.patch("utils.llm.client", return_value=mock_client)
+    client_instance = mocker.patch("src.utils.llm.client", return_value=mock_client)
     client_instance.chat.return_value = create_mock_chat_response(content_response, tool_calls=tool_calls)
 
     (name, params) = call_model_with_tools(system_prompt, user_prompt, tools)
@@ -82,9 +81,9 @@ def test_call_model_with_tools_returns_tuple(mocker):
 
 
 def test_call_model_with_tools_calls_client_chat(mocker):
-    client_instance = mocker.patch("utils.llm.client", return_value=mock_client)
+    client_instance = mocker.patch("src.utils.llm.client", return_value=mock_client)
     client_instance.chat.return_value = create_mock_chat_response(content_response, tool_calls=tool_calls)
-    config_instance = mocker.patch("utils.llm.config", return_value=mock_config)
+    config_instance = mocker.patch("src.utils.llm.config", return_value=mock_config)
 
     call_model_with_tools(system_prompt, user_prompt, tools)
